requirements.txt
pyTelegramBotAPI==4.14.0
mysql-connector-python==8.2.0
requests==2.31.0
python-dotenv==1.0.0
